Move CATSOutput dataclass into it's own module
Hard to avoid a circular import otherwise.
